chore(tests): splt API tests
