+++ test/api/test_declaration.py
@@ -41,22 +41,6 @@ async def test_cannot_get_declaration_without_token(client):
     assert resp.status == 401
 
 
-async def test_request_token(client, monkeypatch):
-    calls = 0
-
-    def mock_send(to, subject, body):
-        assert to == "foo@bar.org"
-        assert "/?token=" in body
-        nonlocal calls
-        calls += 1
-
-    client.logout()
-    monkeypatch.setattr("egapro.emails.send", mock_send)
-    resp = await client.post("/token", body={"email": "foo@bar.org"})
-    assert resp.status == 204
-    assert calls == 1
-
-
 async def test_invalid_siren_should_raise(client, body):
     resp = await client.put("/declaration/111111111/2019", body=body)
     assert resp.status == 422
@@ -216,235 +200,6 @@ async def test_confirmed_declaration_should_raise_if_missing_id(
 async def test_with_unknown_siren_or_year(client):
     resp = await client.get("/declaration/514027945/2019")
     assert resp.status == 404
-
-
-async def test_start_new_simulation(client, body):
-    resp = await client.post("/simulation", body=body)
-    assert resp.status == 200
-    data = json.loads(resp.body)
-    assert "id" in data
-    assert await db.simulation.get(data["id"])
-
-
-async def test_get_simulation(client):
-    uid = await db.simulation.create({"foo": "bar"})
-    resp = await client.get(f"/simulation/{uid}")
-    assert resp.status == 200
-    data = json.loads(resp.body)
-    assert "last_modified" in data
-    del data["last_modified"]
-    assert data == {
-        "data": {"foo": "bar"},
-        "id": uid,
-    }
-
-
-async def test_basic_simulation_should_save_data(client):
-    resp = await client.put(
-        "/simulation/12345678-1234-5678-9012-123456789012",
-        body={"data": {"foo": "bar"}},
-    )
-    assert resp.status == 200
-    data = json.loads(resp.body)
-    assert "last_modified" in data
-    del data["last_modified"]
-    assert data == {
-        "data": {"foo": "bar"},
-        "id": "12345678-1234-5678-9012-123456789012",
-    }
-
-
-async def test_start_new_simulation_send_email_if_given(client, monkeypatch):
-    calls = 0
-    email_body = ""
-
-    def mock_send(to, subject, txt, html=None):
-        assert to == "foo@bar.org"
-        nonlocal calls
-        nonlocal email_body
-        email_body = txt
-        calls += 1
-
-    monkeypatch.setattr("egapro.emails.send", mock_send)
-    resp = await client.post("/simulation", body={"foo": "bar"})
-    assert resp.status == 200
-    assert not calls
-    resp = await client.post(
-        "/simulation",
-        body={"data": {"déclarant": {"email": "foo@bar.org"}}},
-    )
-    assert resp.status == 200
-    data = json.loads(resp.body)
-    assert data["id"] in email_body
-
-
-async def test_put_simulation_should_redirect_to_declaration_if_validated(client):
-    resp = await client.put(
-        "/simulation/12345678-1234-5678-9012-123456789012",
-        body={
-            "data": {
-                "déclarant": {"email": "foo@bar.org"},
-                "status": "valid",
-                "entreprise": {"siren": "12345678"},
-                "déclaration": {"année_indicateurs": 2019},
-            },
-        },
-    )
-    assert resp.status == 307
-    assert resp.headers["Location"] == "/declaration/12345678/2019"
-    # Simulation should have been saved too
-    assert await db.simulation.get("12345678-1234-5678-9012-123456789012")
-
-
-async def test_get_simulation_should_redirect_to_declaration_if_validated(client):
-    uid = await db.simulation.create(
-        {
-            "déclarant": {"email": "foo@bar.org"},
-            "status": "valid",
-            "entreprise": {"siren": "12345678"},
-            "déclaration": {"année_indicateurs": 2019},
-        }
-    )
-    resp = await client.get(f"/simulation/{uid}")
-    assert resp.status == 307
-    assert resp.headers["Location"] == "/declaration/12345678/2019"
-    assert "api-key" in resp.cookies
-
-
-async def test_stats_endpoint(client):
-    rows = [
-        ("12345671", "1000 et plus"),
-        ("12345672", "1000 et plus"),
-        ("12345673", "251 à 999"),
-        ("12345674", "50 à 250"),
-        ("12345675", "50 à 250"),
-        ("12345676", "50 à 250"),
-    ]
-    for siren, tranche in rows:
-        await db.declaration.put(
-            siren,
-            2019,
-            "foo@bar.org",
-            {"informations": {"trancheEffectifs": tranche}},
-        )
-    resp = await client.get("/stats")
-    assert resp.status == 200
-    assert json.loads(resp.body) == {"1000 et plus": 2, "251 à 999": 1, "50 à 250": 3}
-
-
-async def test_send_code_endpoint(client, monkeypatch, body):
-    calls = 0
-    email_body = ""
-    recipient = None
-
-    def mock_send(to, subject, txt, html=None):
-        assert to == "foo@bar.org"
-        nonlocal calls
-        nonlocal email_body
-        nonlocal recipient
-        email_body = txt
-        recipient = to
-        calls += 1
-
-    monkeypatch.setattr("egapro.emails.send", mock_send)
-
-    # Invalid UUID
-    resp = await client.post("/simulation/unknown/send-code", body=body)
-    assert resp.status == 400
-    assert json.loads(resp.body) == {
-        "error": 'Invalid data: invalid input syntax for type uuid: "unknown"'
-    }
-    assert not calls
-
-    # Not found UUID
-    resp = await client.post(
-        "/simulation/12345678-1234-5678-9012-123456789012/send-code",
-        body=body,
-    )
-    assert resp.status == 404
-    assert not calls
-
-    # Create simulation
-    uid = await db.simulation.create(
-        {
-            "déclaration": {"formValidated": "Valid"},
-            "entreprise": {"siren": "12345678"},
-            "informations": {"année_indicateurs": 2019},
-        }
-    )
-
-    # Missing email
-    resp = await client.post(f"/simulation/{uid}/send-code", body=body)
-    assert resp.status == 400
-    assert json.loads(resp.body) == {"error": "Missing `email` key"}
-    assert not calls
-
-    # Valid request.
-    resp = await client.post(
-        f"/simulation/{uid}/send-code", body={"email": "foo@bar.org"}
-    )
-    assert resp.status == 204
-    assert uid in email_body
-    assert recipient == "foo@bar.org"
-
-
-async def test_search_endpoint(client):
-    await db.declaration.put(
-        "12345671",
-        2019,
-        "foo@bar.org",
-        {
-            "déclaration": {"index": 95, "année_indicateurs": 2019},
-            "id": "12345678-1234-5678-9012-123456789013",
-            "entreprise": {
-                "raison_sociale": "Bio c Bon",
-                "effectif": {"tranche": "1000:"},
-            },
-        },
-    )
-    # Only public data should be returned: https://github.com/SocialGouv/egapro-api/pull/14
-    await db.declaration.put(
-        "12345673",
-        2019,
-        "foo@bar.org",
-        {
-            "déclaration": {"index": 95, "année_indicateurs": 2019},
-            "id": "12345678-1234-5678-9012-123456789015",
-            "entreprise": {
-                "raison_sociale": "Bio c Bon moins de 1000",
-                "effectif": {"tranche": "251:999"},
-            },
-        },
-    )
-    resp = await client.get("/search?q=bio")
-    assert resp.status == 200
-    assert json.loads(resp.body) == {
-        "data": [
-            {
-                "declaration": {"noteIndex": 95},
-                "id": "12345678-1234-5678-9012-123456789013",
-                "informations": {"anneeDeclaration": 2019},
-                "informationsEntreprise": {"nomEntreprise": "Bio c Bon"},
-            },
-        ],
-        "total": 1,
-    }
-    resp = await client.get("/search?q=bio&limit=1")
-    assert resp.status == 200
-    assert len(json.loads(resp.body)["data"]) == 1
-
-
-async def test_config_endpoint(client):
-    resp = await client.get("/config")
-    assert resp.status == 200
-    assert list(json.loads(resp.body).keys()) == [
-        "YEARS",
-        "EFFECTIFS",
-        "DEPARTEMENTS",
-        "REGIONS",
-        "REGIONS_TO_DEPARTEMENTS",
-    ]
-    assert json.loads(resp.body)["YEARS"] == [2018, 2019]
 
 
 async def test_declare_with_flat_data(client, body):
