+++ test/test_import_solen.py
